Safari-609.1.11.1/LayoutTests/fast/images/slower-decoding-than-animation-image-expected.txt

Ensure the image frame is drawn when it finishes decoding even if it takes more than the previous frame duration.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.imageFrameIndex(image) is 1
PASS successfullyParsed is true

TEST COMPLETE
